# api/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/passengers": {
      "post": {
        "operationId": "PassengersController_create",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePassengerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassengerResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure."
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a passenger in the current tenant.",
        "tags": [
          "Passengers"
        ]
      },
      "get": {
        "operationId": "PassengersController_list",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 25,
              "example": 25,
              "type": "number"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "mila",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPassengersResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure."
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List passengers in the current tenant.",
        "tags": [
          "Passengers"
        ]
      }
    },
    "/passengers/search": {
      "get": {
        "operationId": "PassengersController_search",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 25,
              "example": 25,
              "type": "number"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "mila",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPassengersResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure."
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Search passengers by first name, last name, phone, or email.",
        "tags": [
          "Passengers"
        ]
      }
    },
    "/passengers/check-duplicates": {
      "post": {
        "operationId": "PassengersController_checkDuplicates",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckPassengerDuplicatesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassengerDuplicatesResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure."
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Find existing passengers that match the supplied full name or phone (diacritic-insensitive).",
        "tags": [
          "Passengers"
        ]
      }
    },
    "/passengers/{id}": {
      "get": {
        "operationId": "PassengersController_getById",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassengerResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          },
          "404": {
            "description": "Passenger not found in current tenant."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get passenger detail by id in the current tenant.",
        "tags": [
          "Passengers"
        ]
      },
      "patch": {
        "operationId": "PassengersController_update",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePassengerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassengerResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure."
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          },
          "404": {
            "description": "Passenger not found in current tenant."
          },
          "409": {
            "description": "Deactivating the passenger would leave active reservations behind.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WouldBreakReservationsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update passenger fields in the current tenant.",
        "tags": [
          "Passengers"
        ]
      },
      "put": {
        "operationId": "PassengersController_replace",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePassengerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassengerResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure."
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          },
          "404": {
            "description": "Passenger not found in current tenant."
          },
          "409": {
            "description": "Deactivating the passenger would leave active reservations behind.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WouldBreakReservationsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Replace passenger fields in the current tenant.",
        "tags": [
          "Passengers"
        ]
      },
      "delete": {
        "operationId": "PassengersController_remove",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassengerResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          },
          "404": {
            "description": "Passenger not found in current tenant."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a passenger in the current tenant.",
        "tags": [
          "Passengers"
        ]
      }
    },
...
  },
  "info": {
    "title": "SURP API",
    "description": "Backend API for SURP, providing tenant-scoped authentication, operational health endpoints, and secure session management for transportation workflows. Protected endpoints require BOTH Authorization: Bearer <accessToken> and X-Tenant-Slug headers (except /platform/* routes), and the tenant must match the token claim.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token issued by /auth/login. Use together with X-Tenant-Slug on protected endpoints."
      }
    },
    "schemas": {
...
      "WouldBreakReservationsDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "WOULD_BREAK_RESERVATIONS"
          },
          "invariant": {
            "type": "string",
            "example": "reservation.seatWithinCapacity",
            "description": "Key of the invariant the change would violate."
          },
          "affectedCount": {
            "type": "number",
            "example": 8,
            "description": "New invariant violations the proposed change would introduce."
          },
          "message": {
            "type": "string",
            "example": "Ova izmena cini 8 rezervacija nevidljivim."
          }
        },
        "required": [
          "code",
          "invariant",
          "affectedCount",
          "message"
        ]
      },
...
      "CreatePassengerDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Mila"
          },
          "lastName": {
            "type": "string",
            "example": "Markovic"
          },
          "phone": {
            "type": "string",
            "example": "+381640000111"
          },
          "email": {
            "type": "string",
            "example": "mila.markovic@demo.local"
          },
          "passengerType": {
            "type": "string",
            "enum": [
              "ADULT",
              "CHILD",
              "STUDENT",
              "SENIOR"
            ],
            "default": "ADULT"
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "default": true
          },
          "notes": {
            "type": "string",
            "example": "VIP customer"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "phone"
        ]
      },
      "PassengerResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "createdById": {
            "type": "object",
            "nullable": true
          },
          "updatedById": {
            "type": "object",
            "nullable": true
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "object",
            "nullable": true
          },
          "passengerType": {
            "type": "string",
            "enum": [
              "ADULT",
              "CHILD",
              "STUDENT",
              "SENIOR"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "notes": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenantId",
          "createdById",
          "updatedById",
          "firstName",
          "lastName",
          "phone",
          "email",
          "passengerType",
          "isActive",
          "notes",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedPassengersResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PassengerResponseDto"
            }
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 25
          },
          "total": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ]
      },
      "CheckPassengerDuplicatesDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Mila"
          },
          "lastName": {
            "type": "string",
            "example": "Marković"
          },
          "phone": {
            "type": "string",
            "example": "+381640000111"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "phone"
        ]
      },
      "PassengerDuplicatesResponseDto": {
        "type": "object",
        "properties": {
          "matches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PassengerResponseDto"
            }
          }
        },
        "required": [
          "matches"
        ]
      },
      "UpdatePassengerDto": {
        "type": "object",
        "properties": {
          "confirmBreakingChange": {
            "type": "boolean",
            "example": false,
            "description": "Confirms a change previously refused with WOULD_BREAK_RESERVATIONS."
          },
          "firstName": {
            "type": "string",
            "example": "Mila"
          },
          "lastName": {
            "type": "string",
            "example": "Markovic"
          },
          "phone": {
            "type": "string",
            "example": "+381640000111"
          },
          "email": {
            "type": "string",
            "example": "mila.markovic@demo.local"
          },
          "passengerType": {
            "type": "string",
            "enum": [
              "ADULT",
              "CHILD",
              "STUDENT",
              "SENIOR"
            ]
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "notes": {
            "type": "string",
            "example": "VIP customer"
          }
        }
      },
...
    }
  }
}